Fixes types
